+++ CHANGELOG.md
@@ -19,6 +19,9 @@
 `reflection` module:
 - `decodeObject()` now auto-validates the object if it implements `Validatable`
 - Wrapped `ObjectEncoding` exceptions into `RuntimeException`
+`network` module:
+- Add `RESTAPIResponse` class to turn EncodableResult into REST API compatible form
+- Add `EncodableResult.toResponse()` method to convert EncodableResult into REST API compatible form
 ### Fixed Bugs
 `core-mit` module:
 - Wrapping of `WrappedException` now returns the original `WrappedException`